Derive NovaInput trait.
