+++ circuits/circuit_input_macro/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2021"
 
 [dependencies]
 serde_json = "1.0.96"
+pasta_curves = "0.5.1"
