+++ circuits/src/nova.rs
@@ -1,3 +1,4 @@
+use circuit_input_macro::NovaInput;
 use nova_scotia::{
     circom::circuit::R1CS, create_public_params, create_recursive_circuit, FileLocation, G2, S1, S2,
 };
@@ -5,11 +6,6 @@ use nova_snark::{traits::Group, CompressedSNARK};
 use pasta_curves::Fq;
 use serde_json::Value;
 use std::{collections::HashMap, path::PathBuf, time::Instant};
-
-pub trait NovaInput {
-    fn initial_inputs(&self) -> Vec<Fq>;
-    fn round_inputs(&self) -> Vec<HashMap<String, Value>>;
-}
 
 pub fn nova(
     iteration_count: usize,
